# docs/swagger.yaml
definitions:
  internal_api.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login request
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      summary: Login
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh request
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration request
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      summary: Register an account
      tags:
      - auth
  /api/v1/courses:
    get:
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      security:
      - BearerAuth: []
      summary: List courses
      tags:
      - courses
    post:
      consumes:
      - application/json
      parameters:
      - description: Course
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      security:
      - BearerAuth: []
      summary: Create course
      tags:
      - courses
  /api/v1/courses/{id}:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete course
      tags:
      - courses
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      security:
      - BearerAuth: []
      summary: Get course
      tags:
      - courses
    put:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Course
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      security:
      - BearerAuth: []
      summary: Update course
      tags:
      - courses
  /api/v1/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      security:
      - BearerAuth: []
      summary: Dashboard statistics
      tags:
      - dashboard
  /api/v1/lessons/{id}/progress:
    post:
      consumes:
      - application/json
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: string
      - description: Progress
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      security:
      - BearerAuth: []
      summary: Report lesson progress
      tags:
      - learning
  /api/v1/resources:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Resource file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      security:
      - BearerAuth: []
      summary: Upload a resource
      tags:
      - resources
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      summary: Service health
      tags:
      - system
  /health/db:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      summary: Database health
      tags:
      - system
swagger: "2.0"